Separate common code for remote side FW upgrade supported optics
