+++ sonic_platform_base/sonic_xcvr/codes/credo/aec_800g.py
@@ -1,8 +1,5 @@
-from ..public.cmis import CmisCodes
+from ..public.cmisTargetFWUpgrade import CmisTargetFWUpgradeCodes
 
-class CmisAec800gCodes(CmisCodes):
-    TARGET_MODE = {
-        0: 'local',
-        1: 'remote-A',
-        2: 'remote-B'
-    }
+class CmisAec800gCodes(CmisTargetFWUpgradeCodes):
+    # Vendor specific implementation to be added here
+    pass
