+++ sonic_platform_base/sonic_xcvr/codes/public/cmisTargetFWUpgrade.py
@@ -1,0 +1,8 @@
+from .cmis import CmisCodes
+
+class CmisTargetFWUpgradeCodes(CmisCodes):
+    TARGET_MODE = {
+        0: 'local',
+        1: 'remote-A',
+        2: 'remote-B'
+    }
